// src/modules/invoice/usecase/find/find.usecase.dto.ts
export interface FindInvoiceUsCaseInputDto {
    id: string;
}

export interface FindInvoiceUsCaseOutputDto {
    id: string;
    name: string;
    document: string;
    address: {
        street: string;
        number: string;
        complement: string;
        city: string;
        state: string;
        zipCode: string;
    };
    items: {
        id: string;
        name: string;
        price: number;
    }[];
    total: number;
    createdAt: Date;
}